requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-huggingface
langchain-community
torch
transformers
accelerate
sentence-transformers